requirements.txt
# Router Communication
librouteros==3.2.0
paramiko==2.11.0

# Data Processing & AI
pandas==2.0.1
numpy==1.24.3
scikit-learn==1.2.2
joblib==1.2.0

# Web Dashboard
Flask==2.3.2
Werkzeug==2.3.4
Jinja2==3.1.2
gunicorn==20.1.0

# Utilities
python-dotenv==1.0.0
schedule==1.2.0
plotly==5.14.1
pytest==7.3.1